+++ src/figaroh/tools/__init__.py
@@ -21,6 +21,7 @@ from . import robotvisualization
 from . import robotcollisions
 from . import robotipopt
 from . import urdf_exporter
+from . import geometric_calibration_export
 from . import export_validation
 from . import report
 from . import identification_report